+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.5.0]
 
+### Added
+
+- Print current front volume into debug monitor when rotating encoder
+
 ### Fixed
 
 - Known bugs on the install script